OSIS)5403 : abstract for load sql
